Fixes tests target not compiling
